refactor: isolate skill eval model
